feat!: rewrite how AmneziaWG obfuscation values are handled
BREAKING CHANGE: removal of `AmneziaSettings` in favor of
`AmneziaWG::V1()` and `AmneziaWG::V2()` enums.

Preview update with `amneziawg-1` and `amneziawg-2` was a mistake, I'm
sorry. I thought it was good, because it didn't break compatibility, but
when I built documentation (with docsrs cfg), the `AmneziaSettings` for
version 1 was missing and it was erroring because of ambiguous name in
globs.
