+++ CHANGELOG.md
@@ -1,14 +1,15 @@
-# v2.0.0 (proposed)
+# v2.0.0
 
 ## Breaking changes
 
-- (propose) Removed `amneziawg` feature
+- Rewritten how AmneziaWG values are handled. Now `amneziawg` feature has both versions of
+  AmneziaWG.
 
 # v1.1.0
 
 ## New features
 
-- Added separate `amneziawg-1` and `amneziawg-2` for AmneziaWG major versions
+- ~~Added separate `amneziawg-1` and `amneziawg-2` for AmneziaWG major versions~~
 - Added support for AmneziaWG 2.0 (by [@Im-a-rabbit](https://github.com/Im-a-rabbit))
 - Added 1.88 as MSRV
 - Added Apache 2.0 license
