+++ src/prelude.rs
@@ -2,3 +2,5 @@
 
 pub use crate::models::*;
 pub use crate::utils::*;
+#[cfg(feature = "amneziawg")]
+pub use crate::amneziawg::*;
