+++ tests/amneziawg/mod.rs
@@ -1,0 +1,4 @@
+#![cfg(feature = "amneziawg")]
+
+mod ver1;
+mod ver2;
