+++ tests/tests.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "amneziawg")]
+mod amneziawg;
+
 mod interface;
 mod peer;
 mod utils;
