+++ tests/utils/mod.rs
@@ -1,4 +1,1 @@
-mod amneziawg1;
-mod amneziawg2;
-
 mod keys;
